+++ server/migrations/20200101130537-add-volume-to-player.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = {
+  up: (queryInterface, Sequelize) => {
+    return queryInterface.addColumn('Players', 'volume', Sequelize.STRING);
+  },
+
+  down: (queryInterface, Sequelize) => {
+    return queryInterface.removeColumn('Players', 'volume');
+  }
+};