+++ server/models/player.js
@@ -3,7 +3,8 @@ module.exports = (sequelize, DataTypes) => {
   const Player = sequelize.define('Player', {
     name: DataTypes.STRING,
     token: DataTypes.STRING,
-    new_remote_code: DataTypes.STRING
+    new_remote_code: DataTypes.STRING,
+    volume: DataTypes.STRING,
   }, {});
   Player.associate = function(models) {
     // associations can be defined here