+++ src/features/noise/noiseSlice.js
@@ -6,15 +6,18 @@ let noisePlayer = null;
 
 const noiseSlice = createSlice({
   name: 'noise',
-  initialState: { noisePlayerInitialized: false, isPlaying: false },
+  initialState: { noisePlayerInitialized: false, isPlaying: false, volume: '-10' },
   reducers: {
     initNoisePlayer(state, action) {
-      noisePlayer = new NoisePlayer(action.payload);
+      noisePlayer = new NoisePlayer(action.payload, state.volume);
       return Object.assign({}, state, { noisePlayerInitialized: true });
     },
     setIsPlaying(state, action) {
       return Object.assign({}, state, { isPlaying: action.payload });
     },
+    setVolume(state, action) {
+      return Object.assign({}, state, { volume: action.payload });
+    }
   }
 });
 
@@ -29,6 +32,17 @@ export const toggleNoise = () => {
   }
 };
 
-export const { initNoisePlayer, setIsPlaying } = noiseSlice.actions;
+export const setNoiseVolume = (volume) => {
+  return (dispatch, getState) => {
+    const { appMode, remote } = getState();
+
+    if (noisePlayer !== null) {
+      noisePlayer.setVolume(volume, appMode, remote.playerId);
+      dispatch(setVolume(volume));
+    }
+  }
+};
+
+export const { initNoisePlayer, setIsPlaying, setVolume } = noiseSlice.actions;
 
 export default noiseSlice.reducer;