+++ src/features/player/playerSlice.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
 import { createSlice } from 'redux-starter-kit';
+import { setNoiseVolume } from '../noise/noiseSlice';
 
 const playerSlice = createSlice({
   name: 'player',
@@ -21,16 +22,17 @@ const playerSlice = createSlice({
 export const { setFetched } = playerSlice.actions;
 
 export function fetchPlayer(playerId) {
-  return (dispatch, getState) => {
+  return (dispatch) => {
     return axios.get(`/api/player/${playerId}`).then(res => {
       dispatch(playerSlice.actions.setPlayer(res.data.player));
       dispatch(setFetched(true));
+      dispatch(setNoiseVolume(res.data.player.volume));
     });
   };
 }
 
 export function createPlayer() {
-  return (dispatch, getState) => {
+  return (dispatch) => {
     return axios.post('/api/player').then(res => {
       dispatch(playerSlice.actions.setPlayer(res.data.player));
     });