+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The client now saves the player/remote setting in the browser.
 * The noise toggle button now shows an animation when the noise is running.
 * Added the flow for authenticating a remote with a specific player.
+* Added a volume slider that is kept in sync between the player and the remotes.
+
 
 ## [0.1.0] - 2019-11-02